More super refactor for material importer
