+++ Engine/ResourceManagement/Importer/TextureImporter.cpp
@@ -2,8 +2,6 @@
 #include "Main/Application.h"
 
 #include <algorithm>
-
-#include <assimp/scene.h>
 #include "Brofiler/Brofiler.h"
 TextureImporter::TextureImporter()
 {
@@ -57,29 +55,6 @@ std::pair<bool, std::string> TextureImporter::Import(const File & file) const
 	}
 	SaveMetaFile(file, output_file);
 	return std::pair<bool, std::string>(true, output_file);
-}
-
-void TextureImporter::ImportMaterialFromMesh(const aiScene* scene, size_t mesh_index, const char* file_path,std::vector<std::string> & loaded_meshes_materials) const
-{
-	int mesh_material_index = scene->mMeshes[mesh_index]->mMaterialIndex;
-	std::string model_base_path = std::string(file_path);
-	aiTextureMapping mapping = aiTextureMapping_UV;
-	for (size_t i = 0; i < AI_TEXTURE_TYPE_MAX; i++)
-	{
-		aiTextureType type = static_cast<aiTextureType>(i);
-		for (size_t j = 0; j < scene->mMaterials[mesh_material_index]->GetTextureCount(type); j++)
-		{
-			aiString file;
-			scene->mMaterials[mesh_material_index]->GetTexture(type, j, &file, &mapping, 0);
-			std::string material_texture = ImportMaterialData(file.data, model_base_path);
-			int texture_type = static_cast<int>(GetTextureTypeFromAssimpType(type));
-			material_texture = std::to_string(texture_type) + ":" + material_texture;
-			if (!material_texture.empty())
-			{
-				loaded_meshes_materials.push_back(material_texture);
-			}
-		}
-	}
 }
 
 std::string TextureImporter::ImportMaterialData(const std::string & material_path, const std::string model_base_path) const
@@ -222,28 +197,6 @@ std::string TextureImporter::GetTextureFileName(std::string texture_file_path) c
 		std::string texture_filename = texture_file_path.substr(found, texture_file_path.length());
 
 		return texture_filename;
-	}
-}
-
-Texture::TextureType TextureImporter::GetTextureTypeFromAssimpType(aiTextureType type) const
-{
-	switch (type)
-	{
-	case aiTextureType_DIFFUSE:
-		return Texture::TextureType::DIFUSSE;
-		break;
-	case aiTextureType_SPECULAR:
-		return Texture::TextureType::SPECULAR;
-		break;
-	case aiTextureType_EMISSIVE:
-		return Texture::TextureType::EMISSIVE;
-		break;
-	case aiTextureType_AMBIENT_OCCLUSION:
-		return Texture::TextureType::OCLUSION;
-		break;
-	default:
-		return Texture::TextureType::UNKNOWN;
-		break;
 	}
 }
